Avoided using of CTimer in CSceneObject and child, in CTestApp and childs (use double mTime variable)

+++ src/sceneobjects/explosion.h
@@ -43,7 +43,7 @@ protected:
 
 private:
     CExplosionParams mParams;
-    float mTime;
+    double mTime;
 };
 
 }// namespace drash
